requirements.txt
llama-index
python-dotenv
langchain 
langchain-community
langchain-openai
openai 
tiktoken 
chromadb
pytest

# pinecone
unstructured==0.7.12 
pinecone-client 
clean-text 
langchain-pinecone
